feat(automation-hub): create separate project for contracts

+++ VerticalWarehouses/Machine Automation Service/Contracts/Hubs/EventArgs/ElevatorPositionChangedEventArgs.cs
@@ -13,7 +13,7 @@
 
         #region Properties
 
-        public int Position { get; }
+        public decimal Position { get; }
 
         #endregion
     }
